flask-blog-posts-main/requirements.txt
Flask~=2.0.2
Flask-SQLAlchemy
flask-login
Flask-WTF
email_validator
Pillow~=8.4.0
WTForms~=3.0.1
python-dotenv
cryptography
gunicorn
psycopg2-binary
